쓰레기통 등록/정보 입력 리팩토링
