+++ src/types/LocationType.ts
@@ -1,4 +1,4 @@
-export type LocationType = {
+export interface LocationType {
   latitude: number;
   longitude: number;
-};
+}
